split encoding tests into v2 and cl files
